requirements.txt
flask
requests
gunicorn
Flask-SQLAlchemy
psycopg2
Flask-Migrate
Flask-Script